Add Magit Configuration
We also explored using `ligature.el` to setup ligature support on
Linux, but ultimately abandoned it since the package is not available
on GNU ELPA (and I don't want to pull in MELPA at the moment). I kept
the ligature configuration around for Berkeley Mono for when I figure
out what I want to do with the package.