Add Support for Local Initialization
We optionally load a `local-init.el` file to allow for local
initialization code to run that we do not want to commit to Git.

Unrelated small change: we specify a tree sitter loading directory in
the `~/.cache/emacs` root, to the same tune as our other directory
customizations where we move files that are not meant to be added to
source control to the `~/.cache` directory.