Keybind Eglot Format
Add a keybind (C-c M-q; as an analogy to M-q) to format the entire
buffer. Note that we use `use-package` (for consistency, and I imaging
that it just Does the Right Thing (TM) with respect to lazy loading)
in order to configure the local keybindings.

Also learned that we don't need to require the whole `use-package` and
can instead just require `use-package-ensure` in order to configure
the "always ensure" setting.