Mardown Mode Remap to GFM
I use GitHub flavoured markdown for everything, so I want to setup the
major mode remapping accordingly. It turns out this needs to happen in
an init and not config, since the remapping needs to be setup
**before** the major loads (and not after). This makes it so the first
load of a mardown file correctly switches to GFM instead of only
working from the second markdown file onwards.