Completely Refactor Emacs Configuration
We now have a monolithic `init.el` instead of spliting it into files.
I noticed that the splitting wasn't so helpful anyway... Maybe in the
future when the config grows, I can think about splitting again
(although, I will try to avoid that situation as much as possible).

One large functional change is that I no longer try to fight Emacs to
get it to put files in '~/.cache/emacs'. It was annoying, and it is
just easier and more practical to just use a '.gitignore' in order to
ignore any files other than my configuration (and since there are so
few - two in fact - it is not a hard list to maintain).